src/subscriber-sdk-direct/requirements.txt
azure-servicebus
azure-identity
aiohttp
python-dotenv
jsons